+++ crystal_generator/__init__.py
@@ -2,7 +2,6 @@ import bpy
 
 from . import crystal_gen_operators, crystal_gen_ui, crystal_gen_utils
 
-
 bl_info = {
     'name' : 'Procedural Crystal Generator',
     'author' : 'Martin Moen',
@@ -10,7 +9,7 @@ bl_info = {
     'location' : 'View3D',
     'category' : 'Add Mesh',
     'blender' : (4, 1, 0),
-    'version' : (0, 1, 1),
+    'version' : (0, 1, 2),
 }
 
 class PCG_PT_MainPanel(bpy.types.Panel):
@@ -18,7 +17,7 @@ class PCG_PT_MainPanel(bpy.types.Panel):
     bl_label = f"Procedural Crystal Generator v. {bl_info['version'][0]}.{bl_info['version'][1]}.{bl_info['version'][2]}"
     bl_space_type = "VIEW_3D"
     bl_region_type = "UI"
-    # bl_category = "PCG"
+    bl_category = "Crystal Generator"
 
     def draw(self, context):
         layout = self.layout
@@ -38,6 +37,7 @@ def register():
 
 
 def unregister():
+
     crystal_gen_ui.unregister()
     crystal_gen_operators.unregister()