+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### ADDED
+- Custom icon
+- Calling create crystal generator through "Add mesh" UI
+
 ### CHANGED
 - Moved props into a pointer group